Settings refactoring adjust & changes
